Put console config under a property in APIConfig
